# docs/swagger.yaml
basePath: /
definitions:
  domain.Comment:
    properties:
      body:
        type: string
      createdDate:
        type: string
      deletedDate:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      postID:
        type: integer
      updatedDate:
        type: string
    type: object
  domain.Post:
    properties:
      body:
        type: string
      createdDate:
        type: string
      deletedDate:
        type: string
      id:
        type: integer
      title:
        type: string
      updatedDate:
        type: string
      userID:
        type: integer
    type: object
  domain.User:
    properties:
      createdDate:
        type: string
      deletedDate:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      updatedDate:
        type: string
    type: object
  requests.CommentRequest:
    properties:
      body:
        example: lorem ipsum
        type: string
    required:
    - body
    type: object
  requests.LoginAuth:
    properties:
      email:
        example: example@email.com
        type: string
      password:
        example: "01234567890"
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  requests.PostRequest:
    properties:
      body:
        example: Lorem ipsum
        type: string
      title:
        example: Lorem ipsum
        type: string
    required:
    - body
    - title
    type: object
  requests.RegisterAuth:
    properties:
      email:
        example: example@email.com
        type: string
      name:
        minLength: 3
        type: string
      password:
        example: "01234567890"
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  response.CommentResponse:
    properties:
      body:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      post_id:
        type: integer
    type: object
  response.LoginResponse:
    properties:
      accessToken:
        type: string
      exp:
        type: integer
      refreshToken:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: REST service for NIX TRAINEE PROGRAM
  title: NIX TRAINEE PROGRAM Demo App
  version: V1.echo
paths:
  /api/v1/comments/comment/{id}:
    get:
      description: Get Comment
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Comment'
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get Comment
      tags:
      - Comments Actions
  /api/v1/comments/delete/{id}:
    delete:
      description: Delete Comment
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete Comment
      tags:
      - Comments Actions
  /api/v1/comments/save:
    post:
      consumes:
      - application/json
      description: Save Comment
      parameters:
      - description: comment info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.CommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.CommentResponse'
        "400":
          description: Bad Request
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Save Comment
      tags:
      - Comments Actions
  /api/v1/comments/update:
    put:
      consumes:
      - application/json
      description: Update Comment
      parameters:
      - description: comment info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.CommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Comment'
        "404":
          description: Not Found
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update Comment
      tags:
      - Comments Actions
  /api/v1/posts/delete/{id}:
    delete:
      description: Delete Post
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete Post
      tags:
      - Posts Actions
  /api/v1/posts/post/{id}:
    get:
      description: Get Post
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Post'
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get Post
      tags:
      - Posts Actions
  /api/v1/posts/save:
    post:
      consumes:
      - application/json
      description: Save Post
      parameters:
      - description: comment info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.PostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Post'
        "400":
          description: Bad Request
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Save Post
      tags:
      - Posts Actions
  /api/v1/posts/update:
    put:
      consumes:
      - application/json
      description: Update Post
      parameters:
      - description: post info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.PostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Post'
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update Post
      tags:
      - Posts Actions
  /login:
    post:
      consumes:
      - application/json
      description: LoginAuth
      parameters:
      - description: users email, users password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.LoginAuth'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.LoginResponse'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
      summary: LoginAuth
      tags:
      - Auth Actions
  /register:
    post:
      consumes:
      - application/json
      description: New user registration
      operationId: user-register
      parameters:
      - description: users email, users password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.RegisterAuth'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema: {}
      summary: Register
      tags:
      - Auth Actions
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"